Rename componet from ShoppingBasket to Header
